requirements.txt
Django
django-rest-framework
pre-commit
